Filter tests using Screenshot annotation
