+++ Corona-Warn-App/src/androidTest/java/de/rki/coronawarnapp/ui/onboarding/OnboardingFragmentTest.kt
@@ -1,9 +1,17 @@
 package de.rki.coronawarnapp.ui.onboarding
 
 import androidx.fragment.app.testing.launchFragment
+import androidx.fragment.app.testing.launchFragmentInContainer
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.action.ViewActions.click
+import androidx.test.espresso.action.ViewActions.scrollTo
+import androidx.test.espresso.assertion.ViewAssertions.matches
+import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
+import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
+import de.rki.coronawarnapp.R
 import io.mockk.MockKAnnotations
 import io.mockk.impl.annotations.MockK
 import org.junit.After
@@ -11,6 +19,8 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import testhelpers.BaseUITest
+import testhelpers.Screenshot
+import tools.fastlane.screengrab.Screengrab
 
 @RunWith(AndroidJUnit4::class)
 class OnboardingFragmentTest : BaseUITest() {
@@ -34,6 +44,18 @@ class OnboardingFragmentTest : BaseUITest() {
     @Test
     fun launch_fragment() {
         launchFragment<OnboardingFragment>()
+    }
+
+    @Screenshot
+    @Test
+    fun capture_screenshot() {
+        launchFragmentInContainer<OnboardingFragment>(themeResId = R.style.AppTheme)
+        // Check any view to make sure screenshot is not blank
+        onView(withId(R.id.onboarding_button_next)).check(matches(isDisplayed()))
+        Screengrab.screenshot(OnboardingFragment::class.simpleName)
+
+        onView(withId(R.id.onboarding_easy_language)).perform(scrollTo(), click())
+        Screengrab.screenshot(OnboardingFragment::class.simpleName.plus("2"))
     }
 }
