+++ Corona-Warn-App/src/androidTest/java/de/rki/coronawarnapp/ui/onboarding/OnboardingPrivacyFragmentTest.kt
@@ -1,9 +1,14 @@
 package de.rki.coronawarnapp.ui.onboarding
 
 import androidx.fragment.app.testing.launchFragment
+import androidx.fragment.app.testing.launchFragmentInContainer
+import androidx.test.espresso.Espresso
+import androidx.test.espresso.assertion.ViewAssertions
+import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
+import de.rki.coronawarnapp.R
 import io.mockk.MockKAnnotations
 import io.mockk.impl.annotations.MockK
 import org.junit.After
@@ -11,6 +16,8 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import testhelpers.BaseUITest
+import testhelpers.Screenshot
+import tools.fastlane.screengrab.Screengrab
 
 @RunWith(AndroidJUnit4::class)
 class OnboardingPrivacyFragmentTest : BaseUITest() {
@@ -34,6 +41,16 @@ class OnboardingPrivacyFragmentTest : BaseUITest() {
     @Test
     fun launch_fragment() {
         launchFragment<OnboardingPrivacyFragment>()
+    }
+
+    @Screenshot
+    @Test
+    fun capture_screenshot() {
+        launchFragmentInContainer<OnboardingPrivacyFragment>(themeResId = R.style.AppTheme)
+        // Check any view to make sure screenshot is not blank
+        Espresso.onView(ViewMatchers.withId(R.id.onboarding_button_next))
+            .check(ViewAssertions.matches(ViewMatchers.isDisplayed()))
+        Screengrab.screenshot(OnboardingPrivacyFragment::class.simpleName)
     }
 }
