+++ Corona-Warn-App/src/androidTest/java/de/rki/coronawarnapp/ui/onboarding/OnboardingTracingFragmentTest.kt
@@ -1,9 +1,15 @@
 package de.rki.coronawarnapp.ui.onboarding
 
 import androidx.fragment.app.testing.launchFragment
+import androidx.fragment.app.testing.launchFragmentInContainer
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.assertion.ViewAssertions.matches
+import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
+import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
+import de.rki.coronawarnapp.R
 import io.mockk.MockKAnnotations
 import io.mockk.impl.annotations.MockK
 import org.junit.After
@@ -11,6 +17,8 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import testhelpers.BaseUITest
+import testhelpers.Screenshot
+import tools.fastlane.screengrab.Screengrab
 
 @RunWith(AndroidJUnit4::class)
 class OnboardingTracingFragmentTest : BaseUITest() {
@@ -34,6 +42,16 @@ class OnboardingTracingFragmentTest : BaseUITest() {
     @Test
     fun launch_fragment() {
         launchFragment<OnboardingTracingFragment>()
+    }
+
+    @Screenshot
+    @Test
+    fun capture_screenshot() {
+        val simpleName = OnboardingTracingFragment::class.simpleName
+        launchFragmentInContainer<OnboardingTracingFragment>(themeResId = R.style.AppTheme)
+        // Check any view to make sure screenshot is not blank
+        onView(withId(R.id.onboarding_button_next)).check(matches(isDisplayed()))
+        Screengrab.screenshot(simpleName)
     }
 }
