--- Corona-Warn-App/src/androidTest/java/de/rki/coronawarnapp/ui/onboarding/screenshot/OnboardingTestFragmentScreenshot.kt
@@ -1,55 +1,0 @@
-package de.rki.coronawarnapp.ui.onboarding.screenshot
-
-import androidx.fragment.app.testing.launchFragmentInContainer
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.withId
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import de.rki.coronawarnapp.R
-import de.rki.coronawarnapp.ui.onboarding.OnboardingTestFragment
-import de.rki.coronawarnapp.ui.onboarding.OnboardingTestViewModel
-import io.mockk.MockKAnnotations
-import io.mockk.impl.annotations.MockK
-import org.junit.After
-import org.junit.Before
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-import testhelpers.BaseUITest
-import testhelpers.Screenshot
-import tools.fastlane.screengrab.Screengrab
-import tools.fastlane.screengrab.locale.LocaleTestRule
-
-@Screenshot
-@RunWith(AndroidJUnit4::class)
-class OnboardingTestFragmentScreenshot : BaseUITest() {
-
-    @MockK lateinit var viewModel: OnboardingTestViewModel
-
-    @Rule
-    @JvmField
-    val localeTestRule = LocaleTestRule()
-
-    @Before
-    fun setup() {
-        MockKAnnotations.init(this, relaxed = true)
-
-        setupMockViewModel(object : OnboardingTestViewModel.Factory {
-            override fun create(): OnboardingTestViewModel = viewModel
-        })
-    }
-
-    @After
-    fun teardown() {
-        clearAllViewModels()
-    }
-
-    @Test
-    fun capture_screenshot() {
-        launchFragmentInContainer<OnboardingTestFragment>(themeResId = R.style.AppTheme)
-        // Check any view to make sure screenshot is not blank
-        onView(withId(R.id.onboarding_button_next)).check(matches(isDisplayed()))
-        Screengrab.screenshot(OnboardingTestFragment::class.simpleName)
-    }
-}
