+++ Corona-Warn-App/src/androidTest/java/testhelpers/Screenshot.kt
@@ -1,5 +1,8 @@
 package testhelpers
 
+/**
+ * Annotation for Screenshots test cases. it help to mark those cases and run them alone
+ */
 @kotlin.annotation.Retention(AnnotationRetention.RUNTIME)
 @Target(
     AnnotationTarget.FUNCTION,
