+++ Corona-Warn-App/src/main/res/layout/include_interoperability.xml
@@ -63,6 +63,7 @@
     </data>
 
     <ScrollView
+        android:id="@+id/interoperability_layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:contentDescription="@{title}">
